% Generated by roxygen2 (4.0.2): do not edit by hand
\name{allmeaningcombinations}
\alias{allmeaningcombinations}
\title{Enumerate meaning combinations.}
\usage{
allmeaningcombinations(dimensionality, uniquelabels = TRUE, offset = 0)
}
\arguments{
\item{dimensionality}{either a) a vector of integers specifying the number
of different possible values for every meaning dimension, or b) a list or
other (potentially ragged) 2-dimensional data structure listing the
possible meaning values for every dimension}

\item{uniquelabels}{logical, determines whether the same integers can be
reused across meaning dimensions or not. When \code{uniquelabels==FALSE},
the resulting matrix will be very reminiscent of tables listing all binary
combinations of factors. Ignored when \code{dimensionality} specifies the
meaning values}

\item{offset}{a constant that is added to all meaning specifiers. Ignored
when \code{dimensionality} specifies the meaning values}
}
\value{
a matrix that has as many columns as there are dimensions, with every
  row specifying one of the possible meaning combinations, and the first
  dimension cycling slowest.
}
\description{
Enumerates all possible combinations of meanings for a meaning space of the given dimensionality.
}
\details{
The resulting matrix can be passed straight on to \code{\link{hammingdists}}.
}
\examples{
allmeaningcombinations(c(2,2))
allmeaningcombinations(c(3,4))
allmeaningcombinations(c(2,2,2,2))
allmeaningcombinations(8) # trivial
allmeaningcombinations(list(shape=c("square", "circle"), color=c("red", "blue")))
}
\seealso{
\code{\link{hammingdists}}
}

