% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix.R
\name{binaryfeaturematrix}
\alias{binaryfeaturematrix}
\title{Convert a categorical meaning matrix to a binary feature matrix.}
\usage{
binaryfeaturematrix(meanings, rownames = NULL)
}
\arguments{
\item{meanings}{a matrix with meaning dimensions along columns and different
meaning combinations along rows (such as created by
\code{\link{enumerate.meaningcombinations}}).}

\item{rownames}{optional character vector of the same length as the number
of rows of \code{meanings}.}
}
\value{
A matrix of \code{TRUE}/\code{FALSE} values.
}
\description{
Convert a categorical meaning matrix to a binary feature matrix.
}
\examples{
enumerate.meaningcombinations(c(2, 2))
binaryfeaturematrix(enumerate.meaningcombinations(c(2, 2)))
}
