% Generated by roxygen2 (4.0.2): do not edit by hand
\name{check.dist}
\alias{check.dist}
\title{Create a symmetric distance matrix.}
\usage{
check.dist(m)
}
\value{
a symmetric matrix with 0s in the diagonal
}
\description{
If \code{m} is a matrix, check whether it is a valid specification of a
distance matrix and return it, making it symmetric if it isn't already.
For all other obect types, try to coerce \code{m} to a \code{dist} object
and return the corresponding distance matrix.
}
\seealso{
\code{\link{stats::dist}}
}

