% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compositionality.R
\name{compositionality}
\alias{compositionality}
\title{An implementation of Spike's measure of compositionality.}
\usage{
compositionality(meaningmatrix, strings)
}
\arguments{
\item{meaningmatrix}{a}

\item{strings}{a}
}
\value{
a list
}
\description{
An implementation of Spike's measure of compositionality.
}
\references{
Spike, M. 2016 \emph{...}. PhD thesis, The University of Edinburgh.
}
