% Generated by roxygen2 (4.0.2): do not edit by hand
\name{conflate.rows}
\alias{conflate.rows}
\title{Aggregate rows and columns of a symmetric (distance) matrix.}
\usage{
conflate.rows(m, groups, discard = FALSE)
}
\arguments{
\item{m}{a symmetric matrix, can be a matrix or \code{dist} object}

\item{groups}{a list of group indices/ids of the same length as the
dimension of the matrix, indicating which rows/columns should be
aggregated over}
}
\value{
a matrix with as many rows and columns as there are unique ids in
  the \code{groups} parameter
}
\description{
Conflates the given matrix by either averaging across or discarding
rows which are marked by the same group index/id.
}

