% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compositionality.R
\name{count.substring.occurrences}
\alias{count.substring.occurrences}
\title{Count occurence of all substrings in a string.}
\usage{
count.substring.occurrences(strings)
}
\arguments{
\item{strings}{a list or vector of character sequences}
}
\value{
a matrix
}
\description{
Returns a matrix with the original strings along rows and all substrings
of those strings along columns. The cell values indicate whether (and how
many times) the substring is contained in each of the strings.
}
\examples{
count.substring.occurrences(c("asd", "asdd"))
}
