% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compositionality.R
\name{enumerate.substrings}
\alias{enumerate.substrings}
\title{Enumerate all substrings of a string.}
\usage{
enumerate.substrings(string)
}
\arguments{
\item{string}{a character string}
}
\value{
a vector containing all substrings of the string (including duplicates)
}
\description{
Enumerate all substrings of a string.
}
\examples{
enumerate.substrings("abccc")
}
