% Generated by roxygen2 (4.0.2): do not edit by hand
\name{hammingdists}
\alias{hammingdists}
\title{Compute all pairwise hamming distances between matrix rows.}
\usage{
hammingdists(meanings)
}
\description{
Returns a distance matrix of all meanings. The meanings argument should be
a matrix with the different dimensions encoded along columns, and all
combinations of meanings specified along rows. The data type of the cells
does not matter since distance is simply based on equality - in fact
specifying a meaning component as NA allows you to ignore that dimension for
the given row/meaning combinations (see examples).
}
\examples{
# example #1: a 2x2 design using strings (the character redundantly
# specifies the dimension for clarity)
hammingdists(matrix(c("a1", "b1", "a1", "b2", "a2", "b1", "a2", "b2"), ncol=2, byrow=TRUE))
# example #2: a 2x3 design using integers for encoding
hammingdists(matrix(c(0,0,0,1,0,2,1,0,1,1,1,2), ncol=2, byrow=TRUE))
# example #3: a 2x2x2 design using factors (ncol is always the number of dimensions)
hammingdists(matrix(c(0,0,0,0,0,1,0,1,0,0,1,1,1,0,0,1,0,1,1,1,0,1,1,1), ncol=3, byrow=TRUE))
# example #4: if some meaning dimension is not relevant for some
# combinations of meanings (e.g. optional arguments), specifying them as NA
# in the matrix will make them not be counted towards the hamming distance!
# in this example the value of the second dimension does not matter (and
# does not count towards the distance) when the the first dimension has
# value '1'
hammingdists(matrix(c(0,0,0,1,1,NA), ncol=2, byrow=TRUE))
}
\seealso{
\code{\link{stats::dist}}
}

