% Generated by roxygen2 (4.0.2): do not edit by hand
\name{mantel.file}
\alias{mantel.file}
\title{Read data from a tab or comma-separated file and run Mantel tests on the data.}
\usage{
mantel.file(filename = NULL, sep = "\\t", header = FALSE,
  stringcolumns = 1, meaningcolumns = -stringcolumns, ...)
}
\arguments{
\item{filename}{csv file to be read - not specifying a file will result in
a file selection dialog being opened}

\item{header}{whether the csv file contains a header line or not}

\item{...}{extra arguments are passed on to \link{\code{mantel.development}}
(and potentially further to \link{\code{plot.mantels}})}
}
\description{
Reads the given .csv file and runs Mantel tests on it, optionally plotting
the results. Plotting is controlled solely through the \code{...} arguments
which are passed on to \link{\code{mantel.development}} and
\link{\code{plot.mantels}} respectively.
If none of the column-arguments are specified it is assumed that the first
column gives the strings and all remaining columns encode the different
meanings. See \link{mantelexample.csv} for a minimal example.
}
\examples{
mantel.file("minimalexample.csv", plot="r")
}
\seealso{
\link{\code{mantel.development}}

\link{\code{plot.mantels}}
}

