% Generated by roxygen2 (4.0.2): do not edit by hand
\name{mantel}
\alias{mantel}
\title{Object class for storing the results of one or more Mantel tests}
\description{
A data frame storing the results of one or more Mantel tests, one test per
row. Multiple \code{mantel} objects can easily be combined by calling
\code{rbind(test1, test2, ...)}
}
\details{
The data frame contains the following columns:
\describe{
   \item{\code{method}}{Character string specifying the type of correlation
         coefficient used (typically "pearson")}
   \item{\code{veridical}}{The correlation coefficient between the distances
         in the two distance matrices}
   \item{\code{sample.size}}{The number of permutations carried out}
   \item{\code{msample}}{A list of correlation coefficients calculated from
         the \code{sample.size} permutations of the input matrices}
   \item{\code{mean}}{Average correlation coefficient produced by the permutations}
   \item{\code{sd}}{Standard deviation of the correlation coefficients}
   \item{\code{z}}{Standard score, \code{(veridical-mean)/sd}}
   \item{\code{ks.level}}{p-value of the Kolmogorov-Smirnov test of
         normality of \code{msample}. If this value is low, \code{sd} and
         \code{z} might not be very meaningful.}
   \item{\code{p.value}}{Empirical p-value computed from the Mantel test:
         let \code{ngreater} be the number of correlation coefficients in
         \code{msample} greater than or equal to \code{veridical}, then
         \code{p.value} is \code{(ngreater+1)(sample.size+1}}
   \item{\code{p.smoothed}}{The theoretical p-value that would correspond
         to the standard \code{z} score as calculated above. If the
         \code{ks.level} is less than was specified as an argument to
         \link{mantel.test}, \code{p.smoothed} will be set to \code{NA}.}
   \item{\code{is.unique.max}}{Logical, \code{TRUE} iff the veridical
         correlation coefficient is greater than any of the coefficients
         calculated for the permutations. If this is true, then
         \code{p.value == 1/(sample.size+1)}}
 }
}

