% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mantel.R
\name{mantel.test}
\alias{mantel.test}
\alias{mantel.test.default}
\alias{mantel.test.formula}
\alias{mantel.test.list}
\title{Perform a Mantel permutation test.}
\usage{
mantel.test(x, y, plot = TRUE, ...)

\method{mantel.test}{default}(x, y, plot = TRUE, method = "pearson",
  trials = 5000, shuffle = shuffle.locations, ...)

\method{mantel.test}{formula}(x, y, plot = TRUE,
  stringdistfun = normalisedlevenshteindists, meaningdistfun = hammingdists,
  ...)

\method{mantel.test}{list}(x, y, plot = TRUE, ...)
}
\arguments{
\item{x}{a formula, distance matrix, or list of distance matrices}

\item{y}{a data frame, distance matrix, or list of matrices of the same length as \code{x}}

\item{plot}{logical: immediately produce a plot of the test results}

\item{...}{extra arguments passed on to the default method}

\item{method}{correlation coefficient to be computed (default: "pearson")}

\item{trials}{maximum number of permutations to be tested}

\item{shuffle}{a function applied to \code{x} to shuffle entries, taking a
matrix and an optional permutation specification as its arguments - you
shouldn't normally have to change this}

\item{stringdistfun}{when \code{x} is a formula: edit distance function used
to compute the distance matrix from the specified string column}

\item{meaningdistfun}{when \code{x} is a formula: meaning distance function
used to compute the distance matrix from the specified meaning columns}
}
\value{
a dataframe of class \code{\link{mantel}} specifying the results of
  the Mantel test(s)
}
\description{
Perform a correlation test between two matrices. The Mantel test is
different from classical correlation tests (such as
\code{\link[stats]{cor.test}}) in that the null distribution of the
correlation coefficient is determined empirically by shuffling the locations
in one of the matrices and calculating the resulting correlations to
generate an empirical null distribution for the given data set.
}
\details{
If the number of possible permutations is reasonably close to
\code{trials}, a deterministic enumeration of all the permutations will be
carried out instead of random sampling.
}
\section{Methods (by class)}{
\itemize{
\item \code{default}: perform Mantel correlation test on two distance
matrices. The distance matrices can either be of type
\code{\link[stats]{dist}}, plain R matrices or any object that can be
interpreted by \code{\link{check.dist}}.

\item \code{formula}: This function can be called with experimental result
data frames, distance matrix calculation is taken care of internally.
\code{x} is a formula of the type \code{s ~ m1 + m2 + ...} where \code{s}
specifies a column of character strings in data frame or matrix \code{y},
while \code{m1} etc. are columns specifying the different meaning
dimensions. To calculate the respective distances, the function
\code{stringdistfun} is applied to the strings, \code{meaningdistfun} to the
meaning columns.

\item \code{list}: when \code{x} is a list of distance matrices, and
\code{y} is either a single distance matrix or a list of distance matrices
the same length as \code{x}. Runs a Mantel test for every pairwise
combination of distance matrices in \code{x} and \code{y} and returns a
\code{\link{mantel}} object with as many rows.
}}

\examples{
# small distance matrix, Mantel test run deterministically
mantel.test(dist(1:7), dist(1:7))

# smallest distance matrix using random permutations (based on default trials=5000)
mantel.test(dist(1:8), dist(1:8))
mantel.test(word ~ Var1 + Var2, cbind(word=c("aa", "ab", "ba", "bb"),
  enumerate.meaningcombinations(c(2, 2))))

# running tests on a list of distance matrices
mantel.test(list(dist(1:16), dist(sample(16:1))),
  hammingdists(enumerate.meaningcombinations(c(2, 2, 2, 2))))
}
\seealso{
\code{\link{plot.mantel}}, \code{\link[stats]{cor}},
  \code{\link{hammingdists}}, \code{\link{normalisedlevenshteindists}},
  \code{\link{orderinsensitivedists}}
}
