% Generated by roxygen2 (4.0.2): do not edit by hand
\name{mcombn}
\alias{mcombn}
\title{Calculate the multinomial coefficient (number of permutations of a multiset)}
\usage{
mcombn(n, ks)
}
\arguments{
\item{n}{}

\item{ks}{vector of positive integers specifying the multiplicities of the
individual elements. '1' elements can be omitted.}
}
\description{
Calculates the number of possible orders in which n elements can be drawn
when some of the individual elements are identical. Pure R implementation
which is fast enough but might fail for large n/k. In this case you probably
want to have a look at \code{multinom} in the \code{multicool} package.
}
\examples{
# only one order for the same sample drawn five times
mcombn(5,5)
# the full 5! possible orders of drawing five unique samples
mcombn(5,rep(1,5))
# sum(ks) should never exceed n, so this will give funny results:
mcombn(5,c(2,4))
}

