% Generated by roxygen2 (4.0.2): do not edit by hand
\name{meaningspace.optionalarguments}
\alias{meaningspace.optionalarguments}
\title{Construct the meaning space matrix for the 'optional arguments' experiment.}
\usage{
meaningspace.optionalarguments(unspecifieddistance)
}
\description{
Constructs a matrix with all meaning combinations for the optional arguments
experiment. the globally shared part of this meaning space design is 2x2x2x2,
but then the upper half of the meaning space this is again combined with
all combinations of an additional 2x2 space (the ground) while the
corresponding lower half is filled with dummy values for the non-specified
ground according to the unspecifieddistance argument which can be 1, 2 or 3.
}
\examples{
meaningspace.optionalarguments(3)
}

