% Generated by roxygen2 (4.0.2): do not edit by hand
\name{multichoose}
\alias{multichoose}
\title{Calculate the multiset coefficient.}
\usage{
multichoose(n, k)
}
\description{
Calculates the number of possible ways to select k out of n elements with
replacement.
}
\seealso{
\url{http://en.wikipedia.org/wiki/Multiset#Counting_multisets}
}

