% Generated by roxygen2 (4.0.2): do not edit by hand
\name{normalisedlevenshteindists}
\alias{normalisedlevenshteindists}
\title{Compute the normalised Levenshtein distances between strings.}
\usage{
normalisedlevenshteindists(strings)
}
\arguments{
\item{strings}{a vector or list of strings}
}
\value{
a distance matrix of normalised Levenshtein distances between the strings
}
\description{
Compute the normalised Levenshtein distances between strings.
}
\examples{
normalisedlevenshteindists(c("abd", "absolute", "asdasd", "casd"))
}
\seealso{
\code{\link{dist}}
}

