% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/page.R
\name{page.compute.exact}
\alias{page.compute.exact}
\title{Calculate exact significance levels of the Page L statistic.}
\usage{
page.compute.exact(k, N, L)
}
\arguments{
\item{k}{number of conditions/generations}

\item{N}{number of replications/chains}

\item{L}{value of the Page L statistic}
}
\value{
the significance level of L for the given k, N
}
\description{
Returns the null probability that the Page statistic with the given k, N is >= L.
}
\examples{
page.compute.exact(6, 4, 322)
}
\seealso{
\code{\link{page.test}}
}
