% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/page.R
\name{page.L}
\alias{page.L}
\title{Calculate Page's L statistic for the given dataset.}
\usage{
page.L(data, ties.method = "average", verbose = TRUE)
}
\arguments{
\item{data}{a matrix or dataframe}

\item{ties.method}{how to resolve tied ranks. Passed on to \code{\link{rank}}, so
can be one of "average" (the default), "first", "last", "random", "max",
"min"}

\item{verbose}{print the final rankings based on which the L statistic is
computed? default: \code{TRUE}}
}
\description{
Calculate Page's L statistic for the given dataset.
}
\seealso{
\code{\link[base]{rank}}
}
