% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/page.R
\name{page.test}
\alias{page.test}
\title{Perform the Page test of monotonicity.}
\usage{
page.test(data, ...)
}
\arguments{
\item{data}{a matrix or dataframe with the different conditions along
columns and the replications along rows. Conversion of the data to ranks
is taken care of internally (see \code{\link{page.L}})}

\item{...}{extra parameters which are passed on to \code{\link{page.L}}}
}
\value{
a list of class \code{pagetest} (and \code{htest}) containing the following fields:
\describe{
   \item{\code{L}}{Value of the \code{L} statistic for the given data set}
   \item{\code{k}}{Number of conditions (columns of the data set)}
   \item{\code{N}}{Number of replications (rows of the data set)}
   \item{\code{p.value}}{Significance level}
   \item{\code{p.type}}{Whether the computed p value is "exact" or "approx"}
}
}
\description{
The Page test tests the given matrix for monotonically increasing ranks
across \code{k} linearly ordered conditions (along columns) based on
\code{N} replications (along rows). To test for monotonically
\emph{decreasing} trends you can either reverse the order of columns, or
simply invert the ranks (i.e. by prepending the dataset with a \code{-}).
}
\details{
Exact p values are computed for \code{k} up to 22, using the pre-computed
null distributions from the \code{pspearman} package. For larger \code{k}, p
values are computed based on a Normal distribution approximation.
}
\examples{
page.test(t(replicate(4, sample(4))))
page.test(t(replicate(4, sample(10))))
}
