% Generated by roxygen2 (4.0.2): do not edit by hand
\name{pages.test.exact.values}
\alias{pages.test.exact.values}
\title{Display the range of exact p values for Page's test provided by this package.}
\usage{
pages.test.exact.values()
}
\description{
Displays a plot of exact p values available to \code{pages.test()} for various k, N
}

