% Generated by roxygen2 (4.0.2): do not edit by hand
\name{pages.test.normal.approx.goodness}
\alias{pages.test.normal.approx.goodness}
\title{Plot the goodness of the normal approximation}
\usage{
pages.test.normal.approx.goodness(p.lvls = c(0.05, 0.01, 0.001))
}
\arguments{
\item{p.lvls}{the significance levels for which the normal approximation
should be compared to the exact values.}
}
\description{
Plot the goodness of the normal approximation
}

