% Generated by roxygen2 (4.0.2): do not edit by hand
\name{pagesL}
\alias{pagesL}
\title{Calculate Page's L statistic for the given dataset.}
\usage{
pagesL(data, ties.method = "average")
}
\arguments{
\item{data}{a matrix or dataframe}

\item{ties}{how to resolve tied ranks. Passed on to \code{\link{rank}}, so can
be one of "average" (the default), "first", "last", "random", "max", "min"}
}
\description{
Calculate Page's L statistic for the given dataset.
}
\seealso{
\code{\link{rank}}
}

