% Generated by roxygen2 (4.0.2): do not edit by hand
\name{plot.mantel}
\alias{plot.mantel}
\title{Plot a Mantel test result.}
\usage{
\method{plot}{mantel}(mantels, plot = "r", xlab = "generation", ...)
}
\arguments{
\item{mantels}{a Mantel test result (a data frame of class 'mantel')}

\item{plot}{the measure to be visualised, currently supported: 'r' and 'z'}

\item{...}{additional parameters to be passed on to the plotting function}
}
\description{
Plots the result of one or more Mantel permutation tests.
}
\examples{
plot(mantel.test(hammingdists(allmeaningcombinations(c(2,2,2,2))), suppressWarnings(as.dist(1:16))))
}
\seealso{
\link{\code{mantel.development}}, \link{\code{plot.default}}
}

