% Generated by roxygen2 (4.0.2): do not edit by hand
\name{plot.mantels}
\alias{plot.mantels}
\title{Plot outcome measures of consecutive Mantel tests.}
\usage{
\method{plot}{mantels}(mantels, plot, xlab = "generation", ...)
}
\arguments{
\item{mantels}{a matrix or data frame of Mantel test results}

\item{plot}{the measure to be visualised, currently supported: 'r' and 'z'}

\item{...}{additional parameters to be passed on to the plotting function}
}
\description{
Plot outcome measures of consecutive Mantel tests.
}
\seealso{
\link{\code{mantel.development}}
}

