% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{pvalue.str}
\alias{pvalue.str}
\title{Convert p-values into strings with inequalities.}
\usage{
pvalue.str(ps, digits = 3, thresholds = c(0.001, 0.01, 0.05))
}
\arguments{
\item{ps}{a vector (or single scalar) of significance levels}

\item{digits}{round maximum digits displayed}

\item{thresholds}{vector of increasing significance levels at which exact
p values should be replaced with "<threshold" labels}
}
\description{
Format p value into a string representation that can be used for publishing.
}
\examples{
pvalue.str(1:8/101)
}
