% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dists.R
\name{segment.counts}
\alias{segment.counts}
\title{Calculate the frequency of individual characters in one or more strings.}
\usage{
segment.counts(x, split = NULL)
}
\arguments{
\item{x}{one or more strings for which segments should be counted}

\item{split}{the boundary character or sequence at which to segment the strings}
}
\value{
a matrix with one row for every string in \code{x}
}
\description{
Calculate the frequency of individual characters in one or more strings.
}
