% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dists.R
\name{segment.string}
\alias{segment.string}
\title{Split one or more strings into their constituent segments.}
\usage{
segment.string(x, split = NULL)
}
\arguments{
\item{x}{one or more strings to be split}

\item{split}{the boundary character or sequence at which to segment the strings}
}
\value{
a list (of the same length as \code{x}) containing character vectors
}
\description{
Vectorisable.
}
