% Generated by roxygen2 (4.0.2): do not edit by hand
\name{shuffle.locations}
\alias{shuffle.locations}
\title{Permute the rows/columns of a matrix.}
\usage{
shuffle.locations(m, perm = sample.int(dim(m)[1]))
}
\arguments{
\item{perm}{vector of indices specifying the new order of rows/columns}

\item{dist}{a matrix with an equal number of rows and columns}
}
\value{
a matrix of the same size as \code{m}
}
\description{
Returns the given matrix with rows and columns permuted in the same order.
}

