% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{temperature.colors}
\alias{temperature.colors}
\title{Create a vector of 'temperature' (from blue over white to red) colors.}
\usage{
temperature.colors(mn, mx = NULL, intensity = 1)
}
\arguments{
\item{mn}{when \code{mx} is not specified: number of colors (>1) in palette}

\item{mx}{maximum}

\item{intensity}{saturation of the most extreme color(s)}
}
\description{
Create a vector of 'temperature' (from blue over white to red) colors.
}
\examples{
image(as.matrix(1:7), z=as.matrix(1:7), col=temperature.colors(7))
image(as.matrix(1:7), z=as.matrix(1:7), col=temperature.colors(-4, 2))
}
\seealso{
\code{\link[grDevices:gray]{grDevices::gray()}}, \code{\link[grDevices:palette]{grDevices::palette()}}, \code{\link[grDevices:rainbow]{grDevices::rainbow()}}
}
