% Generated by roxygen2 (4.0.2): do not edit by hand
\name{mantel.development}
\alias{mantel.development}
\title{Run consecutive Mantel tests on a set of data.}
\usage{
mantel.development(meanings, strings, test.args = NULL, ..., plot = NULL)
}
\arguments{
\item{meanings}{a matrix specifying all meaning combinations, as described in \link{\code{hammingdists}}}

\item{strings}{either a vector of strings, or a matrix of strings per
generation, organised along rows. The result of \code{read.table} can be
passed directly (or \code{t(read.table("..."))} if strings from the same
generation are organised in the same column rather than the same row).}

\item{test.args}{a list of named arguments passed on to \link{\code{mantel.test}}}

\item{...}{extra arguments passed on to the plotting function - here you
might want to specify parameters like \code{ylim}..}

\item{plot}{specifies which mantel test outcome should be plotted (if any):
currently supported are \code{"r"} and \code{"z"}.}
}
\value{
a dataframe of class \code{\link{mantel}} specifying the results of
  the Mantel test(s)
}
\description{
Performs multiple Mantel tests and returns their results in a matrix,
optionally visualising part of the data in a plot. This function can be
called with experimental data, distance matrix calculation is taken care
of internally.
}
\examples{
mantel.development(allmeaningcombinations(c(2,2)), c("asd", "asdf", "", "f"), plot="msample")
m <- matrix(c("sadasd", "iuerwh", "sdfgkj", "uofidsgf", "asd", "asdf", "", "f"), nrow=2, byrow=T)
mantel.development(allmeaningcombinations(c(2,2)), m, plot="r")
mantel.development(allmeaningcombinations(c(2,2)), m, plot="z")
mantel.development(allmeaningcombinations(c(2,2)), read.table(system.file("stringsonlyexample.csv", package="mantel")), plot="r")
}
\seealso{
\link{\code{hammingdists}}

\link{\code{mantel.test}}
}

