% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mantel.R
\name{mantel}
\alias{mantel}
\alias{plot.mantel}
\title{Object class for storing the results of one or more Mantel tests}
\usage{
\method{plot}{mantel}(x, xlab = "generation", ...)
}
\arguments{
\item{x}{a Mantel test result (a data frame of class \code{\link{mantel}})}

\item{xlab}{x axis label when plotting the result of several tests}

\item{...}{additional parameters to be passed on to \code{\link{plot}}}
}
\description{
A data frame storing the results of one or more Mantel tests, one test per
row. Multiple \code{mantel} objects can easily be combined by calling
\code{rbind(test1, test2, ...)}

Plots the result of one or more Mantel permutation tests.
}
\details{
The data frame contains the following columns:
\describe{
   \item{\code{method}}{Character string specifying the type of correlation
         coefficient used ("pearson", "kendall" or "spearman")}
   \item{\code{statistic}}{The veridical correlation coefficient between
         the entries in the two distance matrices}
   \item{\code{sample.size}}{The number of permutations carried out}
   \item{\code{rsample}}{A list of correlation coefficients calculated from
         the \code{sample.size} permutations of the input matrices}
   \item{\code{mean}}{Average correlation coefficient produced by the permutations}
   \item{\code{sd}}{Standard deviation of the sampled correlation coefficients}
   \item{\code{p.value}}{Empirical p-value computed from the Mantel
         test: let \code{ngreater} be the number of correlation coefficients
         in \code{rsample} greater than or equal to \code{statistic}, then
         \code{p.value} is \code{(ngreater+1)/(sample.size+1}}
   \item{\code{p.approx}}{The theoretical p-value that would correspond
         to the standard \code{z} score as calculated above.}
   \item{\code{is.unique.max}}{Logical, \code{TRUE} iff the veridical
         correlation coefficient is greater than any of the coefficients
         calculated for the permutations. If this is true, then
         \code{p.value == 1/(sample.size+1)}}
 }

If the veridical r is plotted in blue it means that it was higher than all
other r's generated by the permutation test.
}
\examples{
plot(mantel.test(hammingdists(enumerate.meaningcombinations(c(2, 2, 2, 2))),
  dist(1:16)))
plot(mantel.test(hammingdists(enumerate.meaningcombinations(c(2, 2, 2, 2))),
  dist(1:16)), plot="sample")
}
\seealso{
\code{\link{mantel.test}}, \code{\link[graphics]{plot}}
}
