% Generated by roxygen2 (4.0.2): do not edit by hand
\name{mantel.test}
\alias{mantel.test}
\title{Perform a Mantel test.}
\usage{
mantel.test(m1, m2, maxtrials = 1000, conflate = FALSE,
  shuffle = shuffle.locations)
}
\arguments{
\item{m1}{a (meaning) distance matrix}

\item{m2}{a (string) distance matrix}

\item{maxtrials}{maximum number of permutations to be tested}

\item{conflate}{if \code{TRUE}, entries with 0 distance in \code{m1} will be
conflated by averaging over the corresponding entries in \code{m2}}

\item{shuffle}{a function applied to \code{m2} to shuffle entries, taking a
matrix and an optional permutation specification as its arguments - you
shouldn't normally have to change this}
}
\value{
a dataframe specifying the results of the Mantel test(s)
}
\description{
Performs a Mantel permutation test on two distance matrices.
}
\examples{
# small distance matrix, Mantel test run deterministically
mantel.test(dist(1:4), dist(1:4))
# smallest distance matrix using random permutations (with maxtrials=1000)
mantel.test(dist(1:7), dist(1:7))
}
\seealso{
\code{\link{hammingdists}}, \code{\link{normalisedlevenshteindists}}
}

