% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dists.R
\name{normalisedlevenshteindists}
\alias{normalisedlevenshteindists}
\title{Compute the normalised Levenshtein distances between strings.}
\usage{
normalisedlevenshteindists(strings)
}
\arguments{
\item{strings}{a vector or list of strings}
}
\value{
A distance matrix specifying all pairwise normalised Levenshtein
distances between the strings.
}
\description{
Compute the normalised Levenshtein distances between strings.
}
\examples{
normalisedlevenshteindists(c("abd", "absolute", "asdasd", "casd"))
}
\seealso{
\code{\link[stats]{dist}}
}
