% Generated by roxygen2 (4.0.2): do not edit by hand
\name{pages.test.exact}
\alias{pages.test.exact}
\title{Calculate exact significance levels for Page's L.}
\usage{
pages.test.exact(k, N, p.lvls = c(0.05, 0.01, 0.001), message.progress = k *
  N >= 30, L = 0)
}
\arguments{
\item{k}{number of conditions/generations}

\item{N}{number of replications/chains}

\item{p.lvls}{vector of p levels for which the exact critical L value is sought}

\item{message.progress}{logical, whether to \code{message()} information on
the progress of the computation to the output}

\item{L}{the lowest L for which the p value should be computed}
}
\description{
Calculate critical values of Page's L for the given significance levels.
}
\details{
If only one p value is supplied, the function will return the full table of
p values for all Ls computed up to that p value, rather than just the
critical L for that particular p.
}
\examples{
pages.test.exact(6, 4, p.lvls=1)
}
\seealso{
page.test
}

