% Generated by roxygen2 (4.0.2): do not edit by hand
\name{pages.test}
\alias{pages.test}
\title{Perform Page's test on the given data set, providing exact p values whre possible.}
\usage{
pages.test(data, ...)
}
\arguments{
\item{data}{a matrix or dataframe with the different conditions along
columns and the replications along rows. Conversion of the data to ranks
is taken care of internally.}

\item{...}{extra parameters which are passed on to \code{\link{pagesL}}}
}
\value{
a list of class \code{PagesTest}
}
\description{
Page's test tests for monotonically increasing ranks across linearly ordered
conditions. To test for monotonically \emph{decreasing} trends you can
either reverse the order of columns, or simply invert the ranks (i.e. by
prepending the dataset with a \code{-}).
}
\examples{
pages.test(t(replicate(4, sample(4))))
pages.test(t(replicate(4, sample(10))))
}

