% Generated by roxygen2 (4.0.2): do not edit by hand
\name{plot.mantel}
\alias{plot.mantel}
\title{Plot a Mantel test result.}
\usage{
\method{plot}{mantel}(mantels, plot = if (nrow(mantels) == 1) "msample" else
  "r", xlab = "generation", ...)
}
\arguments{
\item{mantels}{a Mantel test result (a data frame of class \code{\link{mantel}})}

\item{plot}{the measure to be visualised, currently supported: "r", "z" and "msample"}

\item{...}{additional parameters to be passed on to the plotting function}
}
\description{
Plots the result of one or more Mantel permutation tests. The exact
visualisation used can be controlled with the \code{plot} parameter, which
currently supports "r", "z" and "msample".
}
\details{
Red error bars or normal distribution fits signify that a Kolmogorov-Smirnov
test of the randomised r's was significantly non-normal at the level
specified by \code{ks.level} in \code{mantel.test}. If the veridical r is
plotted in blue it means that it was higher than all other r's generated by
the permutation test.

See \code{\link{mantel.development}} for more interesting example plots.
}
\examples{
plot(mantel.test(hammingdists(allmeaningcombinations(c(2,2,2,2))), suppressWarnings(as.dist(1:16))))
plot(mantel.test(hammingdists(allmeaningcombinations(c(2,2,2,2))), suppressWarnings(as.dist(1:16))), plot="z")
}
\seealso{
\link{\code{mantel.development}}, \link{\code{mantel.test}}, \link{\code{plot.default}}
}

