% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dists.R
\name{shuffle.locations}
\alias{shuffle.locations}
\title{Permute the rows and columns of a square matrix.}
\usage{
shuffle.locations(m, perm = sample.int(dim(m)[1]))
}
\arguments{
\item{m}{a matrix with an equal number of rows and columns}

\item{perm}{vector of indices specifying the new order of rows/columns}
}
\value{
a matrix of the same size as \code{m}
}
\description{
Returns the given matrix with rows and columns permuted in the same order.
}
